HR_Backend_FastAPI/requirements.txt
# FastAPI and Web Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database - MongoDB (includes GridFS for file storage)
pymongo>=4.6.0
motor>=3.3.2

# Validation and Serialization (Pydantic v2)
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-core>=2.14.0

# Environment Variables
python-dotenv>=1.0.0

# File Upload
python-multipart>=0.0.6

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1

# PDF Parsing (for resume extraction)
PyPDF2>=3.0.1
pdfplumber>=0.10.3

# Document Parsing
python-docx>=1.1.0

# HTTP Requests
httpx>=0.25.2
requests>=2.31.0

# Utilities
email-validator>=2.1.0
python-dateutil>=2.8.2

# Optional: AI/NLP (uncomment when ready to integrate)
# openai>=1.3.0
# langchain>=0.1.0
# tiktoken>=0.5.2
